Move IpRange types to omicron_common.
