+++ common/src/address.rs
@@ -7,11 +7,11 @@
 //! This addressing functionality is shared by both initialization services
 //! and Nexus, who need to agree upon addressing schemes.
 
-use crate::api::external::Ipv6Net;
+use crate::api::external::{self, Error, Ipv6Net};
 use ipnetwork::Ipv6Network;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
-use std::net::{Ipv6Addr, SocketAddrV6};
+use std::net::{IpAddr, Ipv4Addr, Ipv6Addr, SocketAddrV6};
 
 pub const AZ_PREFIX: u8 = 48;
 pub const RACK_PREFIX: u8 = 56;
@@ -177,6 +177,187 @@ pub fn get_64_subnet(
     Ipv6Subnet::<SLED_PREFIX>::new(Ipv6Addr::from(rack_network))
 }
 
+/// An IP Range is a contiguous range of IP addresses, usually within an IP
+/// Pool.
+///
+/// The first address in the range is guaranteed to be no greater than the last
+/// address.
+#[derive(Clone, Copy, Debug, PartialEq, Deserialize, Serialize)]
+#[serde(untagged)]
+pub enum IpRange {
+    V4(Ipv4Range),
+    V6(Ipv6Range),
+}
+
+// NOTE: We don't derive JsonSchema. That's intended so that we can use an
+// untagged enum for `IpRange`, and use this method to annotate schemars output
+// for client-generators (e.g., progenitor) to use in generating a better
+// client.
+impl JsonSchema for IpRange {
+    fn schema_name() -> String {
+        "IpRange".to_string()
+    }
+
+    fn json_schema(
+        gen: &mut schemars::gen::SchemaGenerator,
+    ) -> schemars::schema::Schema {
+        schemars::schema::SchemaObject {
+            subschemas: Some(Box::new(schemars::schema::SubschemaValidation {
+                one_of: Some(vec![
+                    external::label_schema(
+                        "v4",
+                        gen.subschema_for::<Ipv4Range>(),
+                    ),
+                    external::label_schema(
+                        "v6",
+                        gen.subschema_for::<Ipv6Range>(),
+                    ),
+                ]),
+                ..Default::default()
+            })),
+            ..Default::default()
+        }
+        .into()
+    }
+}
+
+impl IpRange {
+    pub fn first_address(&self) -> IpAddr {
+        match self {
+            IpRange::V4(inner) => IpAddr::from(inner.first),
+            IpRange::V6(inner) => IpAddr::from(inner.first),
+        }
+    }
+
+    pub fn last_address(&self) -> IpAddr {
+        match self {
+            IpRange::V4(inner) => IpAddr::from(inner.last),
+            IpRange::V6(inner) => IpAddr::from(inner.last),
+        }
+    }
+}
+
+impl From<IpAddr> for IpRange {
+    fn from(addr: IpAddr) -> Self {
+        match addr {
+            IpAddr::V4(addr) => IpRange::V4(Ipv4Range::from(addr)),
+            IpAddr::V6(addr) => IpRange::V6(Ipv6Range::from(addr)),
+        }
+    }
+}
+
+impl TryFrom<(Ipv4Addr, Ipv4Addr)> for IpRange {
+    type Error = String;
+
+    fn try_from(pair: (Ipv4Addr, Ipv4Addr)) -> Result<Self, Self::Error> {
+        Ipv4Range::new(pair.0, pair.1).map(IpRange::V4)
+    }
+}
+
+impl TryFrom<(Ipv6Addr, Ipv6Addr)> for IpRange {
+    type Error = String;
+
+    fn try_from(pair: (Ipv6Addr, Ipv6Addr)) -> Result<Self, Self::Error> {
+        Ipv6Range::new(pair.0, pair.1).map(IpRange::V6)
+    }
+}
+
+/// A non-decreasing IPv4 address range, inclusive of both ends.
+///
+/// The first address must be less than or equal to the last address.
+#[derive(Clone, Copy, Debug, PartialEq, Deserialize, Serialize, JsonSchema)]
+#[serde(try_from = "AnyIpv4Range")]
+pub struct Ipv4Range {
+    first: Ipv4Addr,
+    last: Ipv4Addr,
+}
+
+impl Ipv4Range {
+    pub fn new(first: Ipv4Addr, last: Ipv4Addr) -> Result<Self, String> {
+        if first <= last {
+            Ok(Self { first, last })
+        } else {
+            Err(String::from("IP address ranges must be non-decreasing"))
+        }
+    }
+
+    pub fn first_address(&self) -> Ipv4Addr {
+        self.first
+    }
+
+    pub fn last_address(&self) -> Ipv4Addr {
+        self.last
+    }
+}
+
+impl From<Ipv4Addr> for Ipv4Range {
+    fn from(addr: Ipv4Addr) -> Self {
+        Self { first: addr, last: addr }
+    }
+}
+
+#[derive(Clone, Copy, Debug, Deserialize)]
+struct AnyIpv4Range {
+    first: Ipv4Addr,
+    last: Ipv4Addr,
+}
+
+impl TryFrom<AnyIpv4Range> for Ipv4Range {
+    type Error = Error;
+    fn try_from(r: AnyIpv4Range) -> Result<Self, Self::Error> {
+        Ipv4Range::new(r.first, r.last)
+            .map_err(|msg| Error::invalid_request(msg.as_str()))
+    }
+}
+
+/// A non-decreasing IPv6 address range, inclusive of both ends.
+///
+/// The first address must be less than or equal to the last address.
+#[derive(Clone, Copy, Debug, PartialEq, Deserialize, Serialize, JsonSchema)]
+#[serde(try_from = "AnyIpv6Range")]
+pub struct Ipv6Range {
+    first: Ipv6Addr,
+    last: Ipv6Addr,
+}
+
+impl Ipv6Range {
+    pub fn new(first: Ipv6Addr, last: Ipv6Addr) -> Result<Self, String> {
+        if first <= last {
+            Ok(Self { first, last })
+        } else {
+            Err(String::from("IP address ranges must be non-decreasing"))
+        }
+    }
+
+    pub fn first_address(&self) -> Ipv6Addr {
+        self.first
+    }
+
+    pub fn last_address(&self) -> Ipv6Addr {
+        self.last
+    }
+}
+
+impl From<Ipv6Addr> for Ipv6Range {
+    fn from(addr: Ipv6Addr) -> Self {
+        Self { first: addr, last: addr }
+    }
+}
+
+#[derive(Clone, Copy, Debug, Deserialize)]
+struct AnyIpv6Range {
+    first: Ipv6Addr,
+    last: Ipv6Addr,
+}
+
+impl TryFrom<AnyIpv6Range> for Ipv6Range {
+    type Error = Error;
+    fn try_from(r: AnyIpv6Range) -> Result<Self, Self::Error> {
+        Ipv6Range::new(r.first, r.last)
+            .map_err(|msg| Error::invalid_request(msg.as_str()))
+    }
+}
+
 #[cfg(test)]
 mod test {
     use super::*;
@@ -227,5 +408,63 @@ mod test {
             "[fd00:1122:3344:0308::1]:12345".parse::<SocketAddrV6>().unwrap(),
             get_sled_address(subnet)
         );
+    }
+
+    #[test]
+    fn test_ip_range_checks_non_decreasing() {
+        let lo = Ipv4Addr::new(10, 0, 0, 1);
+        let hi = Ipv4Addr::new(10, 0, 0, 3);
+        assert!(Ipv4Range::new(lo, hi).is_ok());
+        assert!(Ipv4Range::new(lo, lo).is_ok());
+        assert!(Ipv4Range::new(hi, lo).is_err());
+
+        let lo = Ipv6Addr::new(0xfd00, 0, 0, 0, 0, 0, 0, 1);
+        let hi = Ipv6Addr::new(0xfd00, 0, 0, 0, 0, 0, 0, 3);
+        assert!(Ipv6Range::new(lo, hi).is_ok());
+        assert!(Ipv6Range::new(lo, lo).is_ok());
+        assert!(Ipv6Range::new(hi, lo).is_err());
+    }
+
+    #[test]
+    fn test_ip_range_enum_deserialization() {
+        let data = r#"{"first": "10.0.0.1", "last": "10.0.0.3"}"#;
+        let expected = IpRange::V4(
+            Ipv4Range::new(
+                Ipv4Addr::new(10, 0, 0, 1),
+                Ipv4Addr::new(10, 0, 0, 3),
+            )
+            .unwrap(),
+        );
+        assert_eq!(expected, serde_json::from_str(data).unwrap());
+
+        let data = r#"{"first": "fd00::", "last": "fd00::3"}"#;
+        let expected = IpRange::V6(
+            Ipv6Range::new(
+                Ipv6Addr::new(0xfd00, 0, 0, 0, 0, 0, 0, 0),
+                Ipv6Addr::new(0xfd00, 0, 0, 0, 0, 0, 0, 3),
+            )
+            .unwrap(),
+        );
+        assert_eq!(expected, serde_json::from_str(data).unwrap());
+
+        let data = r#"{"first": "fd00::3", "last": "fd00::"}"#;
+        assert!(
+            serde_json::from_str::<IpRange>(data).is_err(),
+            "Expected an error deserializing an IP range with first address \
+            greater than last address",
+        );
+    }
+
+    #[test]
+    fn test_ip_range_try_from() {
+        let lo = Ipv4Addr::new(10, 0, 0, 1);
+        let hi = Ipv4Addr::new(10, 0, 0, 3);
+        assert!(IpRange::try_from((lo, hi)).is_ok());
+        assert!(IpRange::try_from((hi, lo)).is_err());
+
+        let lo = Ipv6Addr::new(0xfd00, 0, 0, 0, 0, 0, 0, 1);
+        let hi = Ipv6Addr::new(0xfd00, 0, 0, 0, 0, 0, 0, 3);
+        assert!(IpRange::try_from((lo, hi)).is_ok());
+        assert!(IpRange::try_from((hi, lo)).is_err());
     }
 }
